+++ apps/blog/app/posts/[id]/page.tsx
@@ -2,30 +2,7 @@ import { format } from "date-fns";
 import MarkdownRenderer from "../../../components/MarkdownRenderer";
 
 import { Badge } from "@raonc/ui/components/badge";
-import { getArticleQueryById } from "../../../constant/hashnodeQuery";
-import { Post } from "../../../type/hashnode";
-
-export const getArticleById = async (id: string) => {
-  const response = await fetch(`https://gql.hashnode.com/`, {
-    cache: "no-cache",
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: `${process.env.HASHNODE_ACCESS_TOKEN}`,
-    },
-    body: JSON.stringify({
-      query: getArticleQueryById(id),
-    }),
-  });
-
-  const {
-    data: {
-      publication: { post },
-    },
-  } = await response.json();
-
-  return post as Post;
-};
+import { getArticleById } from "../../../utils/hashnode";
 
 export default async function PostPage({ params }: { params: { id: string } }) {
   const { title, publishedAt, tags, content, readTimeInMinutes } =
